Refactor: extract semantic-parity report builders into Diagnostics module
Slice 3 of decomposition epic #242. Pure behavior-preserving move of the
semantic-parity report builders (the html_semantic_parity_* findings:
landmark count comparison, navigation_menu_missing,
navigation_item_count_mismatch, and the source-vs-block nav/landmark parity
counting) out of HtmlTransformer into a dedicated
Diagnostics/SemanticParityReporter module.

The new module uses the shared DomHelpersTrait (slice 2) and receives the
parsed source body, generated blocks, and source provenance as explicit
parameters rather than reaching into transformer state. It carries a Runtime
(required by trait helpers) mirroring the transformer's own. The small
hasAncestorTag / hasSourceNavigationSignal / safeNavigationUrl helpers remain
in HtmlTransformer for non-parity callers and are duplicated in the reporter
so it needs no transformer reference (DomHelpersTrait is untouched per slice
constraints).

No logic changes, no new/renamed findings, no changed counting. Parity
128->128, canonical identical, full composer test green. HtmlTransformer
6475 -> 5706 lines (-769).

Refs #242
